+++ spec/dummy/app/controllers/my_posts_controller.rb
@@ -1,0 +1,3 @@
+class MyPostsController < ApplicationController
+  include RestfulJson::DefaultController
+end