+++ spec/dummy/app/models/barfoo.rb
@@ -1,2 +1,3 @@
 class Barfoo < ActiveRecord::Base
+  include AbleToFailOnPurpose
 end