+++ spec/dummy/app/models/foobar.rb
@@ -1,2 +1,3 @@
 class Foobar < ActiveRecord::Base
+  include AbleToFailOnPurpose
 end