+++ spec/dummy/app/models/my_post.rb
@@ -1,0 +1,3 @@
+class MyPost < ActiveRecord::Base
+  include AbleToFailOnPurpose
+end