+++ spec/dummy/app/permitters/post_permitter.rb
@@ -1,0 +1,3 @@
+class MyPostPermitter < ApplicationPermitter
+  permit :name, :title, :content
+end